requirements.txt
numpy
pandas
scipy
scikit-learn
faiss-cpu
gradio
rank-bm25
sentence-transformers
pyarrow
